Improve server performance

+++ lib/server/server_ifs.hpp
@@ -9,7 +9,7 @@ namespace genome {
         std::ifstream m_ifs;
         long m_pos_data_beg;
     protected:
-        void get_counts(const std::vector<size_t>& hashes, std::vector<size_t>& counts) override;
+        void get_counts(const std::vector<size_t>& hashes, std::vector<uint16_t>& counts) override;
     public:
         explicit server_ifs(const boost::filesystem::path& path);
     };
