+++ lib/server/server_mmap.hpp
@@ -8,7 +8,7 @@ namespace genome {
     private:
         byte* m_data;
     protected:
-        void get_counts(const std::vector<size_t>& hashes, std::vector<size_t>& counts) override;
+        void get_counts(const std::vector<size_t>& hashes, std::vector<uint16_t>& counts) override;
     public:
         explicit server_mmap(const boost::filesystem::path& path);
     };
